requirements.txt
---

## 📦 **requirements.txt**
```txt
fastapi
uvicorn
httpx
python-dotenv
xmltodict
pandas
geopy
